requirements.txt
requests
supabase
google-generativeai
python-dotenv
flask
gunicorn